Separate pack/unpack into separate library.
Pack/unpack will now use common pack/unpack library definitions.  This new
library also improves validation, fixing a number of outstanding issues.

Resolves: LE-4708 LE-6312 LE-6314 LE-6316

+++ framework/c/src/safeRef.c
@@ -37,7 +37,7 @@ static const char ModuleName[] = "ref";
 //--------------------------------------------------------------------------------------------------
 typedef struct le_ref_Map
 {
-    ssize_t             nextRefNum;     ///< The next Safe Reference value to be assigned.
+    uint32_t             nextRefNum;     ///< The next Safe Reference value to be assigned.
 
     le_hashmap_Ref_t    referenceMap;    ///< HashMap of Mapping objects.
